+++ src/model/model.hh
@@ -5,13 +5,13 @@
 template <int size>
 struct InferenceData {
 
-    std::array<float, size> strategy_prior0;
-    std::array<float, size> strategy_prior1;
-    float  value_estimate0 = .5f;
-    float  value_estimate1 = .5f;
+    std::array<double, size> strategy_prior0;
+    std::array<double, size> strategy_prior1;
+    double  value_estimate0 = .5f;
+    double  value_estimate1 = .5f;
 
     InferenceData () {}
-    InferenceData (float value_estimate0, float value_estimate1) :
+    InferenceData (double value_estimate0, double value_estimate1) :
     strategy_prior0(strategy_prior0), strategy_prior1(strategy_prior1), value_estimate0(value_estimate0), value_estimate1(value_estimate1) {}
 
 };
@@ -35,7 +35,7 @@ public:
     InferenceData<size> inference (State<size>& state) {
         PairActions<size> pair;
         InferenceData<size> inference_data;
-        float u = state.rollout();
+        double u = state.rollout();
         // fill w/ uniform /.... or dont
         inference_data.value_estimate0 = u;
         inference_data.value_estimate1 = 1-u;
